python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zinb_graphical_model"
version = "0.1.1"
description = "ZINB Graphical Model with PyTorch/Pyro using pseudo-likelihood inference"
readme = "README.md"
license = {text = "GPL-3.0"}
authors = [
    {name = "GW McElfresh", email = "mcelfreshgw@gmail.com"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "torch>=2.6.0",
    "pyro-ppl>=1.8.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/GWMcElfresh/PreGraphModeling"
Repository = "https://github.com/GWMcElfresh/PreGraphModeling"

[tool.setuptools.packages.find]
where = ["."]
include = ["zinb_graphical_model*"]
